Added StartingHostState
* Modified OfflineState and ConnectingState so that client and host flows are similar
